refactor(v4): give the utils a folder and a barrel of their own
`src/utils.ts` sat beside the `src/utils/` directory the maths merge created,
which resolves but reads as a mistake. Everything is in the folder now:

  utils.ts  ->  utils/strings.ts     kebabCase, pascalCase
            ->  utils/selectors.ts   selectorFor
                utils/maths.ts       unchanged
                utils/index.ts       the barrel

`selectorFor` is split out rather than filed under strings because it is not a
spelling — it encodes the attribute contract, which is why `data-component` is
the attribute and `~=` is the matcher. The spec splits the same way, which is
what takes the file count to 25.

Utils are **not** re-exported from `src/index.ts`, and `kebabCase`/
`pascalCase` are dropped from it. The root barrel is the framework's surface —
components, the registry, the scheduler, the services — and a `clamp` beside a
`Base` says nothing true about either. Nothing in the repo imported them from
there; `demo/` takes only `registerComponents`.

Internal imports stay deep, so a module's dependencies are legible and
tree-shaking has nothing to unpick. The barrel is the entry point for a
consumer, and `migration/index.ts` keeps re-exporting the maths on its own
line rather than widening to the whole barrel.
