+++ packages/v4/src/utils/selectors.spec.ts
@@ -1,13 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { kebabCase, selectorFor } from './utils.js';
-
-describe('kebabCase', () => {
-  it('converts PascalCase and camelCase', () => {
-    expect(kebabCase('SliderDragStart')).toBe('slider-drag-start');
-    expect(kebabCase('fetchAfter')).toBe('fetch-after');
-    expect(kebabCase('click')).toBe('click');
-  });
-});
+import { selectorFor } from './selectors.js';
 
 describe('selectorFor', () => {
   it('matches whitespace-separated data-component tokens', () => {
